First Draft of Warp 'warp.cs' created, which is a monobehavior which calls 'OnTriggerStay2D.' class warp has a variable, 'openness,' which indicates if the warp is 'open' (true) or 'closed.' (false)
warp.cs needs debugging. The intent is for a message to be posted to the
console when 'openness' is marked as true, the player is colliding with
'entrance,' and the key 'e' is pressed.

A new object, 'entrance,' was added to the project for debugging
'warp.cs.'

The player layer level was changed to '1' so it is always drawn over the
warp and other objects.

A 2D box collider was added to 'entrance' and 'player' to realize the
warp design.